Fix the behaviour of the ROOT NamedKey of a hive
The ROOT NamedKey of a hive should not have a named part in the path.
This will interfere when constructing a full path for a key when the hive
is mapped on some subkey of another hive.

NamedKey.path() for the ROOT key failed as it did not take into
consideration that it did not have any parents.